analyzer/requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
playwright==1.45.1
opencv-python-headless==4.10.0.84
pillow==10.4.0
